It is now almost done.
And it doesn't look to bad either.

the Home tab shows a table with all the URLs saved in the database.
The Add URL does what the tab suggests.
The Admin tab, if you have not entered an Admin pass before, it will ask you to enter one. When entering a correct pass, you will be granted access to the admin_panel.php. Where you can delete URLs. The only thing I haven't done, is make it so you can update the URLs in the database.